[GREEN] implement create transaksi method

+++ src/main/java/jagongadpro/penjualanpembelian/dto/KeranjangDto.java
@@ -17,6 +17,6 @@ public class KeranjangDto {
 
     private Map<String, Integer> items = new HashMap<>();
 
-    private double totalPrice;
+    private int totalPrice;
 
 }
